Minor changes to XFractint source to clean up Gentoo compile. Other minor cleanup of buffer overflow conditions. Added -no-pie to Makefile.
git-svn-id: https://svn.fractint.net/trunk/fractint@1749 cff58d23-3d7a-4071-9076-c60e8c9d8568

+++ common/fractalb.c
@@ -23,7 +23,7 @@ fractal routines.
 
 
 int bf_math = 0;
-double b_const;
+LDBL b_const;
 
 #if (_MSC_VER >= 700)
 #pragma code_seg ("bigsetup_text")     /* place following in an overlay */
